Move folder src/windows/launcher -> src/launcher
