+++ src/launcher/actions/releaseNotesDialogActions.js
