+++ src/launcher/components/AppIcon.jsx
