+++ src/launcher/components/AppItem.jsx
