+++ src/launcher/components/AppManagementView.jsx
