+++ src/launcher/components/ProxyErrorDialog.jsx
