+++ src/launcher/components/ReleaseNotesDialogView.jsx
