+++ src/launcher/components/Root.jsx
@@ -35,40 +35,38 @@
  */
 
 import React from 'react';
-import PropTypes from 'prop-types';
 
-import ConfirmationDialog from '../../../components/ConfirmationDialog';
+import Nav from 'react-bootstrap/Nav';
+import Tab from 'react-bootstrap/Tab';
 
-const ConfirmLaunchDialog = ({
-    isVisible,
-    text,
-    app,
-    onConfirm,
-    onCancel,
-}) => (
-    <ConfirmationDialog
-        isVisible={isVisible}
-        title="Version problem"
-        text={text}
-        okButtonText="Launch anyway"
-        cancelButtonText="Cancel"
-        onOk={() => onConfirm(app)}
-        onCancel={onCancel}
-    />
+import AppManagementContainer from '../containers/AppManagementContainer';
+import ErrorDialogContainer from '../containers/ErrorDialogContainer';
+import SettingsContainer from '../containers/SettingsContainer';
+import UpdateAvailableContainer from '../containers/UpdateAvailableContainer';
+import UpdateProgressContainer from '../containers/UpdateProgressContainer';
+import ConfirmLaunchContainer from '../containers/ConfirmLaunchContainer';
+import ProxyLoginContainer from '../containers/ProxyLoginContainer';
+import ProxyErrorContainer from '../containers/ProxyErrorContainer';
+import Logo from '../../components/Logo';
+
+export default () => (
+    <>
+        <Tab.Container id="launcher" defaultActiveKey="apps">
+            <Nav>
+                <Nav.Link accessKey="1" eventKey="apps">apps</Nav.Link>
+                <Nav.Link accessKey="2" eventKey="settings">settings</Nav.Link>
+                <Logo />
+            </Nav>
+            <Tab.Content>
+                <Tab.Pane eventKey="apps"><AppManagementContainer /></Tab.Pane>
+                <Tab.Pane eventKey="settings"><SettingsContainer /></Tab.Pane>
+            </Tab.Content>
+        </Tab.Container>
+        <ErrorDialogContainer />
+        <UpdateAvailableContainer />
+        <UpdateProgressContainer />
+        <ConfirmLaunchContainer />
+        <ProxyLoginContainer />
+        <ProxyErrorContainer />
+    </>
 );
-
-ConfirmLaunchDialog.propTypes = {
-    isVisible: PropTypes.bool.isRequired,
-    text: PropTypes.string.isRequired,
-    app: PropTypes.shape({
-        name: PropTypes.string.isRequired,
-    }),
-    onConfirm: PropTypes.func.isRequired,
-    onCancel: PropTypes.func.isRequired,
-};
-
-ConfirmLaunchDialog.defaultProps = {
-    app: null,
-};
-
-export default ConfirmLaunchDialog;
