+++ src/launcher/components/SettingsView.jsx
