+++ src/launcher/components/__tests__/AppManagementView-test.jsx
