+++ src/launcher/components/__tests__/SettingsView-test.jsx
