+++ src/launcher/components/__tests__/UpdateProgressDialog-test.jsx
