+++ src/launcher/containers/AppManagementContainer.js
@@ -1,4 +1,4 @@
-/* Copyright (c) 2015 - 2018, Nordic Semiconductor ASA
+/* Copyright (c) 2015 - 2019, Nordic Semiconductor ASA
  *
  * All rights reserved.
  *
@@ -34,37 +34,45 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import React from 'react';
-import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+import AppManagementView from '../components/AppManagementView';
+import * as AppsActions from '../actions/appsActions';
+import { openUrlInDefaultBrowser } from '../../util/fileUtil';
+import * as DesktopShortcutActions from '../actions/desktopShortcutActions';
+import * as ReleaseNotes from '../actions/releaseNotesDialogActions';
 
-import ConfirmationDialog from '../../../components/ConfirmationDialog';
+function mapStateToProps(state) {
+    const {
+        apps: {
+            localApps, officialApps,
+            installingAppName, removingAppName, upgradingAppName,
+        },
+    } = state;
 
-const ConfirmRemoveSourceDialog = ({
-    isVisible,
-    source,
-    onConfirm,
-    onCancel,
-}) => (
-    <ConfirmationDialog
-        isVisible={isVisible}
-        title="Remove app source"
-        text={`Are you sure to remove "${source}" source along with any apps installed from it?`}
-        okButtonText="Yes, remove"
-        cancelButtonText="Cancel"
-        onOk={() => onConfirm(source)}
-        onCancel={onCancel}
-    />
-);
+    return {
+        apps: localApps.concat(officialApps),
+        installingAppName,
+        removingAppName,
+        upgradingAppName,
+        isProcessing: !!installingAppName || !!upgradingAppName || !!removingAppName,
+    };
+}
 
-ConfirmRemoveSourceDialog.propTypes = {
-    isVisible: PropTypes.bool.isRequired,
-    source: PropTypes.string,
-    onConfirm: PropTypes.func.isRequired,
-    onCancel: PropTypes.func.isRequired,
-};
+function mapDispatchToProps(dispatch) {
+    return {
+        onInstall: (name, source) => dispatch(
+            AppsActions.installOfficialApp(name, source),
+        ),
+        onRemove: (name, source) => dispatch(AppsActions.removeOfficialApp(name, source)),
+        onReadMore: homepage => openUrlInDefaultBrowser(homepage),
+        // Launcher actions
+        onAppSelected: app => dispatch(AppsActions.checkEngineAndLaunch(app)),
+        onCreateShortcut: app => dispatch(DesktopShortcutActions.createShortcut(app)),
+        onShowReleaseNotes: appid => dispatch(ReleaseNotes.show(appid)),
+    };
+}
 
-ConfirmRemoveSourceDialog.defaultProps = {
-    source: null,
-};
-
-export default ConfirmRemoveSourceDialog;
+export default connect(
+    mapStateToProps,
+    mapDispatchToProps,
+)(AppManagementView);
