+++ src/launcher/containers/ConfirmLaunchContainer.js
