+++ src/launcher/containers/ConfirmRemoveSourceDialog.js
