+++ src/launcher/containers/ErrorDialogContainer.js
@@ -34,55 +34,26 @@
  * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import 'core-js/es7';
-import 'regenerator-runtime/runtime';
+import { connect } from 'react-redux';
+import ErrorDialog from '../../components/ErrorDialog';
+import * as ErrorDialogActions from '../../actions/errorDialogActions';
 
-import React from 'react';
-import { render } from 'react-dom';
-import { remote } from 'electron';
-import isDev from 'electron-is-dev';
-import RootContainer from './containers/RootContainer';
-import configureStore from '../../store/configureStore';
-import rootReducer from './reducers';
-import * as AppsActions from './actions/appsActions';
-import * as AutoUpdateActions from './actions/autoUpdateActions';
-import * as ProxyActions from './actions/proxyActions';
-import '../../../resources/css/launcher.scss';
+function mapStateToProps(state) {
+    const { errorDialog } = state;
 
-const config = remote.require('./main/config');
-const settings = remote.require('./main/settings');
-const net = remote.require('./main/net');
-
-const store = configureStore(rootReducer);
-const rootElement = React.createElement(RootContainer, { store });
-
-const shouldCheckForUpdatesAtStartup = settings.get('shouldCheckForUpdatesAtStartup');
-
-function downloadLatestAppInfo() {
-    if (shouldCheckForUpdatesAtStartup !== false && !config.isSkipUpdateApps()) {
-        return store.dispatch(AppsActions.downloadLatestAppInfo());
-    }
-    return Promise.resolve();
+    return {
+        messages: errorDialog.messages,
+        isVisible: errorDialog.isVisible,
+    };
 }
 
-function checkForCoreUpdates() {
-    if (shouldCheckForUpdatesAtStartup !== false && !config.isSkipUpdateCore() && !isDev) {
-        return store.dispatch(AutoUpdateActions.checkForCoreUpdates());
-    }
-    return Promise.resolve();
+function mapDispatchToProps(dispatch) {
+    return {
+        onClose: () => dispatch(ErrorDialogActions.hideDialog()),
+    };
 }
 
-net.registerProxyLoginHandler((authInfo, callback) => {
-    store.dispatch(ProxyActions.authenticate(authInfo))
-        .then(credentials => {
-            const { username, password } = credentials;
-            return callback(username, password);
-        });
-});
-
-render(rootElement, document.getElementById('webapp'), async () => {
-    await store.dispatch(AppsActions.loadLocalApps());
-    await store.dispatch(AppsActions.loadOfficialApps());
-    await downloadLatestAppInfo();
-    await checkForCoreUpdates();
-});
+export default connect(
+    mapStateToProps,
+    mapDispatchToProps,
+)(ErrorDialog);
