+++ src/launcher/containers/ReleaseNotesDialogContainer.js
