+++ src/launcher/containers/SettingsContainer.js
