+++ src/launcher/containers/UpdateProgressContainer.js
