+++ src/launcher/reducers/__tests__/appsReducer-test.js
