+++ src/launcher/reducers/__tests__/autoUpdateReducer-test.js
