+++ src/launcher/reducers/__tests__/proxyReducer-test.js
