+++ src/launcher/reducers/__tests__/settingsReducer-test.js
