+++ src/launcher/reducers/appsReducer.js
