+++ src/launcher/reducers/autoUpdateReducer.js
