+++ src/launcher/reducers/proxyReducer.js
