+++ src/launcher/reducers/releaseNotesDialogReducer.js
