+++ src/launcher/reducers/settingsReducer.js
